refactor out get_zenml_config_dir
