+++ src/zenml/orchestrators/base_orchestrator.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Any, Optional
 from zenml.core.base_component import BaseComponent
 from zenml.core.component_factory import orchestrator_store_factory
 from zenml.enums import OrchestratorTypes
-from zenml.utils.path_utils import get_zenml_config_dir
+from zenml.io.fileio import get_zenml_config_dir
 
 if TYPE_CHECKING:
     from zenml.pipelines.base_pipeline import BasePipeline
