+++ src/zenml/utils/path_utils.py
@@ -14,13 +14,9 @@
 """File utilities"""
 # TODO: [TFX] [LOW] Unnecessary dependency here
 
-import os
-from pathlib import Path
 
 from tfx.dsl.io.filesystem import PathType
 
-from zenml.core.constants import ZENML_DIR_NAME
-from zenml.io import fileio
 from zenml.logger import get_logger
 
 logger = get_logger(__name__)
@@ -426,19 +422,19 @@ _REMOTE_FS_PREFIX = ["gs://", "hdfs://", "s3://"]
 #     return get_zenml_dir(str(Path(path).parent))
 
 
-def get_zenml_config_dir(path: str = os.getcwd()) -> str:
-    """Recursive function to find the zenml config starting from path.
+# def get_zenml_config_dir(path: str = os.getcwd()) -> str:
+#     """Recursive function to find the zenml config starting from path.
 
-    Args:
-        path (Default value = os.getcwd()): Path to check.
+#     Args:
+#         path (Default value = os.getcwd()): Path to check.
 
-    Returns:
-        The full path with the resolved zenml directory.
+#     Returns:
+#         The full path with the resolved zenml directory.
 
-    Raises:
-        InitializationException if directory not found until root of OS.
-    """
-    return os.path.join(fileio.get_zenml_dir(str(Path(path))), ZENML_DIR_NAME)
+#     Raises:
+#         InitializationException if directory not found until root of OS.
+#     """
+#     return os.path.join(fileio.get_zenml_dir(str(Path(path))), ZENML_DIR_NAME)
 
 
 def convert_to_str(path: PathType) -> str:
